requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
sqlalchemy==2.0.44
python-jose==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
email-validator==2.3.0
python-dotenv==1.2.1
